// src/rest/model/weekDeliveryPredictionDays.ts
/**
 * Pakkasmarja REST API
 * REST API for Pakkasmarja
 *
 * OpenAPI spec version: 3.0.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface WeekDeliveryPredictionDays {
    monday: boolean  | null;
    tuesday: boolean  | null;
    wednesday: boolean  | null;
    thursday: boolean  | null;
    friday: boolean  | null;
    saturday: boolean  | null;
    sunday: boolean  | null;
}

export interface WeekDeliveryPredictionDaysOpt {
    monday?: boolean;
    tuesday?: boolean;
    wednesday?: boolean;
    thursday?: boolean;
    friday?: boolean;
    saturday?: boolean;
    sunday?: boolean;
}